requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==2.0.12
h11==0.14.0
idna==3.8
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
py==1.11.0
PySocks==1.7.1
pytest==6.2.5
requests==2.26.0
selenium==4.5.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
toml==0.10.2
trio==0.26.2
trio-websocket==0.11.1
urllib3==1.26.20
wsproto==1.2.0
